LayoutTests/platform/blackberry/editing/text-iterator/findString-markers.html

<!DOCTYPE html>
<html>
<head>
<style>
.plaintext {
    border: 2px solid red;
    padding: 12px;
    font-size: 24px;
}
</style>
<script src="../../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<pre id="description"></pre>
<pre id="console"></pre>
<div id="container">
</div>

<script>

if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

function createPlainTextElement(id, text, parent) {
    var e = document.createElement('div');
    e.className = 'plaintext';
    e.id = id;
    e.innerText = text;

    parent.appendChild(e);
    return e;
}

function occurrences(string, substring) {
    var n=0;
    var pos=0;

    while (true) {
        pos = string.toLowerCase().indexOf(substring.toLowerCase(),pos);
        if (pos!=-1) {
            ++n;
            pos+=substring.length;
        } else
            break;
    }
    return n;
}

function checkTextMatchMarker(index) {
    if (index > 26) {
        layoutTestController.notifyDone();
        return;
    }
    var str = 'abcdefghijklmnopqrstuvwxyz.';
    var char = str.charAt(index);
    layoutTestController.findString(str.charAt(index), []);
    // Start the function later to allow marking process to finish.
    setTimeout(function() {
        var count = occurrences(text, char);
        // The number of markers should match to the occurrences of the charactor.
        shouldBe('internals.markerCountForNode(e.firstChild, "textmatch")', count.toString());
        var searchStartIndex = 0;
        // Get the char from text and compare with the marker range content.
        for (var i = 0; i < count ; ++i) {
            var range = internals.markerRangeForNode(e.firstChild, "textmatch", i);
            var pos = text.toLowerCase().indexOf(char.toLowerCase(), searchStartIndex);
            searchStartIndex = pos + 1;
            var expectStr = text.charAt(pos);
            shouldBe('"' + range + '"', '"' + expectStr + '"');
        }
        checkTextMatchMarker(++index);
    }, 850);
}

var container = document.getElementById('container');

var id = 'id1'
var text = 'The quick brown fox jumps over the lazy dog.';
var e = createPlainTextElement(id, text, container);

if (window.internals && window.layoutTestController)
    checkTextMatchMarker(0);

var successfullyParsed = true;
</script>
<script src="../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
